+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/config/DiscordConfig.kt
@@ -1,9 +1,0 @@
-package io.github.sophon.discord.config
-
-import kotlinx.serialization.Serializable
-
-//TODO: ConfigRepo
-@Serializable
-data class DiscordConfig(
-    val discordBotApiKey: String
-)