+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/config/DiscordConfig.kt
@@ -1,0 +1,9 @@
+package io.github.sophon.discord.feat.config
+
+import kotlinx.serialization.Serializable
+
+//TODO: ConfigRepo
+@Serializable
+data class DiscordConfig(
+    val discordBotApiKey: String
+)