+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/ewgf/usecase/GetRecentMatchesUseCase.kt
@@ -1,0 +1,17 @@
+package io.github.sophon.discord.feat.ewgf.usecase
+
+import io.github.sophon.EwgfClient
+import io.github.sophon.core.domain.Result
+import io.github.sophon.core.domain.mapError
+import io.github.sophon.discord.feat.core.domain.model.BotError
+import io.github.sophon.discord.feat.core.domain.toDomainError
+import io.github.sophon.domain.model.BattleSet
+
+internal class GetRecentMatchesUseCase(
+    private val client: EwgfClient,
+) {
+    suspend fun invoke(discordId: String): Result<List<BattleSet>, BotError> {
+        return client.downloadBattleData(discordId)
+            .mapError { it.toDomainError() }
+    }
+}