+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/ewgf/usecase/RegisterPlayerUseCase.kt
@@ -1,0 +1,18 @@
+package io.github.sophon.discord.feat.ewgf.usecase
+
+import io.github.sophon.EwgfClient
+import io.github.sophon.core.domain.EmptyResult
+import io.github.sophon.core.domain.mapError
+import io.github.sophon.discord.feat.core.domain.model.BotError
+import io.github.sophon.discord.feat.core.domain.toDomainError
+import io.github.sophon.domain.model.Player
+
+internal class RegisterPlayerUseCase(
+    private val client: EwgfClient,
+) {
+    suspend fun invoke(discordId: String, polarisId: String): EmptyResult<BotError> {
+        val player = Player(polarisId = polarisId, discordId =  discordId)
+        return client.registerPlayer(player)
+            .mapError { it.toDomainError() }
+    }
+}