+++ bot/discord/src/jvmTest/kotlin/io/github/sophon/botdiscord/featureRegistry/core/usecase/CreateJoinEmbedButtonUseCaseTest.kt
@@ -1,5 +1,0 @@
-package io.github.sophon.botdiscord.featureRegistry.core.usecase
-
-//TODO: implement test
-//steam://joinlobby/586140/109775241137042824/76561198443042808
-//https://Sophon.github.io/lobby.html?target=steam://joinlobby/586140/109775241137042824/76561198443042808